Améliorations majeures : nouveaux types MIDI, mode pulse press/release, Note balayage avec filtrage
- Ajout support Pitch Bend, Aftertouch, Program Change, Clock, Tap Tempo
- Mode bouton pulse : option press/release pour timing
- Note (balayage) : filtre médian + hystérésis, auto-off 16-bit
- Correction OSC address personnalisée par pin
- Suppression correcte des pins depuis NVS
- Réduction spam console (OSCQueue debug)
- Ajout spécification Clock MIDI complète (doc)
- Support btnPulseTiming dans interface web et backend

+++ src/BluetoothManager.h
@@ -42,7 +42,7 @@ public:
     void sendNoteOff(uint8_t channel, uint8_t note, uint8_t velocity);
     void sendControlChange(uint8_t channel, uint8_t control, uint8_t value);
     void sendProgramChange(uint8_t channel, uint8_t program);
-    void sendPitchBend(uint8_t channel, uint16_t bend);
+    void sendPitchBend(uint8_t channel, int bend);
     
     // État de connexion
     bool isConnected() const;
